Removed part of test that was for Zope 2.12 or lower.
We were testing that we could ask a folder its object values or ids
after we migrated its btree to None.
In Zope 2.12 this returned an empty list.
In Zope 2.13 and higher this returns either an AttributeError
or a TypeError.
We only support 2.13, so it is not really interesting to check this.