Removed ancient migration code for upgrading BTree folder from Plone 3 to 4.
The code, or at least the tests, are incompatible with the latest `zope.interface`.
See https://github.com/plone/plone.app.folder/issues/24
and the remarks in this coredev commit that updated to zope.interface 5.1.1:
https://github.com/plone/buildout.coredev/commit/daefdc903c29041f4b7315a4e8a55f8c59ce5c97

Current plone.app.folder is only for Plone 5.2, so the code is not needed.
The migrate-btrees view is referenced nowhere in the code.